refactor(list-group): add src and types folders
