+++ packages/list-group/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from './ListGroup';
-export * from './ListGroup';
+export { default } from './types/ListGroup';
+export * from './types/ListGroup';
